fix #649 (problem with show.std)
If standardized coefficients are displayed (show.std = TRUE), the model is refitted. For mixed models this can result in different p-values and t-values as compared to the unstandardized coefficients. At the moment this can result in a mismatch between the t-value (coming from the standardized model) and the p-value (coming from the unstandardized model). The fix removes the t-value for standardized coefficients before merging with the unstandardized model. Thus, the t-value and p-value always refer to the unstandardized model.